+++ docs/CHANGELOG.rst
@@ -4,9 +4,9 @@ Change Log
 
 All notable changes to this project are documented in this file.
 
-==========
-Unreleased
-==========
+===================
+22.2.0 - 2025-07-02
+===================
 
 Added
 -----